+++ docs/changelog.md
@@ -12,6 +12,7 @@ pages](https://github.com/dalibo/ldap2pg/pulls?utf8=%E2%9C%93&q=is%3Apr%20is%3Am
 - Search for `/etc/openldap/ldap.conf` too.
 - Protect role name when reassigning objects.
 - Apply `LDAPCONF` `LDAPRC` from environment.
+- Version config file. This will help manage breaking changes in ldap2pg.
 
 
 # ldap2pg 5.5